Color Changes
Finally some progress to show for. I was able to change the color of the
background based on which arrow key is pressed (Up, down, left, right)
and it changes back to default once released. Now its just a matter of
applying this to a moving a character around the screen instead of
changed the color of the background

Interestingly enough. There is a command to do something while a key is
being held. In which case we could do something interesting with
charging shields/weapons, giving awards to the player based on if he
completed the whole level without letting go of the moveForward key,
etc. I think we could have some fun with this.